Move favourite config mapper
