+++ src/frontend/packages/core/src/shared/components/favorites-entity-list/favorites-entity-list.component.ts
@@ -2,8 +2,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { combineLatest, Observable, ReplaySubject, Subject } from 'rxjs';
 import { distinctUntilChanged, map, scan, startWith } from 'rxjs/operators';
 
+import { FavoritesConfigMapper, IFavoriteTypes } from '../../../../../store/src/favorite-config-mapper';
 import { IFavoriteEntity } from '../../../../../store/src/user-favorite-manager';
-import { FavoritesConfigMapper, IFavoriteTypes } from '../favorites-meta-card/favorite-config-mapper';
 
 
 @Component({
