+++ src/frontend/packages/store/src/user-favorite-helpers.ts
@@ -1,6 +1,6 @@
-import { FavoritesConfigMapper } from '../../core/src/shared/components/favorites-meta-card/favorite-config-mapper';
 import { entityCatalog } from './entity-catalog/entity-catalog';
 import { IEntityMetadata } from './entity-catalog/entity-catalog.types';
+import { FavoritesConfigMapper } from './favorite-config-mapper';
 import { IFavoriteMetadata, UserFavorite } from './types/user-favorites.types';
 
 export function isEndpointTypeFavorite(favorite: UserFavorite<IFavoriteMetadata>) {
